requirements.txt
discord.py==2.6.4
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.10.11
openai==2.7.1
schedule==1.2.0
